backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Compress==1.15
Flask-Limiter==3.5.0
pydantic==2.7.1
pydantic-core==2.18.2
boto3==1.34.14
numpy==1.26.3
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
typing_extensions>=4.8.0
